Add editions support to simp-metadata commandlets
* Move edition handling code out of simp/media/engine.rb since it's no longer needed there
* Add Simp::Metadata.uri to wrap URI since URI does not handle git ssh urls well.
* Create writable? parameter for sources, marking bootstrap_source has unwritable.
* Add editions to clone and delete commands.
* Make -w option take multiple metadata sources and urls, to support enterprise edition
* Always write to all writable sources when creating or deleting a release.
* Remove hard coded download logic from Simp::Metadata::Source and use Simp::Metadata.download_component
* Add separate directories when writing out releases, so that metadata repos are more organized

Implements commercial/enterprise-meta#227

+++ spec/simp/media/type/internet_spec.rb
@@ -2,15 +2,28 @@ require 'spec_helper'
 require 'simp/media'
 
 describe Simp::Media::Type::Internet do
-
-  it "should run without errors" do
-    tempdir = Dir.mktmpdir("simp-media-rspec")
-    engine = Simp::Media::Engine.new({"version" => "test-stub", "edition" => "community", "output_type" => "tar", "output" => "#{tempdir}/test.tgz"})
-    expect(engine.loaded?).to match(true)
-    engine.run
-    engine.cleanup()
-    engine = nil
-    FileUtils.rmtree(tempdir)
+  describe "with edition == 'community'" do
+    it "should run without errors" do
+      tempdir = Dir.mktmpdir("simp-media-rspec")
+      engine = Simp::Media::Engine.new({"version" => "test-stub", "edition" => "community", "output_type" => "tar", "output" => "#{tempdir}/test.tgz"})
+      expect(engine.loaded?).to match(true)
+      engine.run
+      engine.cleanup()
+      engine = nil
+      FileUtils.rmtree(tempdir)
+    end
   end
-
+  if (ENV.fetch("SIMP_LICENSE_KEY", nil) != nil)
+    describe "with edition == 'enterprise'" do
+      it "should run without errors" do
+        tempdir = Dir.mktmpdir("simp-media-rspec")
+        engine = Simp::Media::Engine.new({"version" => "test-stub", "edition" => "enterprise", "output_type" => "tar", "output" => "#{tempdir}/test.tgz"})
+        expect(engine.loaded?).to match(true)
+        engine.run
+        engine.cleanup()
+        engine = nil
+        FileUtils.rmtree(tempdir)
+      end
+    end
+  end
 end
